+++ projects/start-os/web/ui/src/app/routes/portal/routes/backups/backups.component.ts
@@ -719,6 +719,9 @@ export default class BackupsComponent implements OnInit {
             })
           }
         }
+        for (const job of this.jobs()) {
+          await this.api.deleteScheduledBackupJob({ id: job.id })
+        }
       }
     } catch (error: any) {
       this.errors.handleError(getErrorMessage(error))